Rename snapshot module to snapshot_map
